refactor hierarchy
