+++ app/src/main/java/com/example/hello/world/app/ui/buttons/model/ButtonModel.kt
@@ -1,0 +1,4 @@
+package com.example.hello.world.app.ui.buttons.model
+
+
+data class ButtonModel(val name: String)
