+++ app/src/main/java/com/example/hello/world/app/ui/home/MainFragment.kt
@@ -1,4 +1,4 @@
-package com.example.hello_world_app.ui.home
+package com.example.hello.world.app.ui.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,8 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.example.hello_world_app.R
-import com.example.hello_world_app.databinding.FragmentMainBinding
+import com.example.hello.world.app.R
+import com.example.hello.world.app.databinding.FragmentMainBinding
 
 
 class MainFragment : Fragment(), View.OnClickListener {
