--- app/src/main/java/com/example/hello_world_app/ui/buttons/model/ButtonModel.kt
@@ -1,4 +1,0 @@
-package com.example.hello_world_app.ui.buttons.model
-
-
-data class ButtonModel(val name: String)
