+++ app/src/main/res/layout/fragment_main.xml
@@ -76,8 +76,7 @@
         app:backgroundTint="@color/light_blue"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/parentLayoutButtonOpenButtons"
-        tools:layout_editor_absoluteY="164dp" />
+        app:layout_constraintTop_toBottomOf="@+id/parentLayoutButtonOpenButtons" />
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/parentLayoutButtonOpenRadio"
